fix: tentativas de link bd
